+++ src/com/javaops/webapp/storage/ArrayStorage.java
@@ -1,9 +1,26 @@
 package com.javaops.webapp.storage;
 
+import com.javaops.webapp.model.Resume;
+
 /**
  * Array based storage for Resumes
  */
 public class ArrayStorage extends AbstractArrayStorage {
+    @Override
+    public void save(Resume r) {
+        if (size == storage.length) {
+            System.out.println(STORAGE_EXCESS_MESSAGE);
+            return;
+        }
+
+        int position = getPosition(r.getUuid());
+        if (position >= 0) {
+            System.out.printf((RESUME_PRESENT_MESSAGE) + "%n", r.getUuid());
+            return;
+        }
+        storage[size++] = r;
+    }
+
     @Override
     public void delete(String uuid) {
         int resumePosition = getPosition(uuid);
